Restore old offset for binary compatibility
